+++ tests/integration/alb/test_alb_update.py
@@ -1,19 +1,19 @@
+from datetime import date
+
 import pytest  # noqa F401
 
 from broker.extensions import db
-from broker.models import ALBServiceInstance
+from broker.models import ALBServiceInstance, Operation
 
 from tests.lib.client import check_last_operation_description
 
 from tests.lib.update import (
-    subtest_update_creates_update_operation,
     subtest_update_creates_private_key_and_csr,
     subtest_gets_new_challenges,
     subtest_update_updates_TXT_records,
     subtest_update_answers_challenges,
     subtest_waits_for_dns_changes,
     subtest_update_retrieves_new_cert,
-    subtest_update_uploads_new_cert,
     subtest_update_marks_update_complete,
     subtest_update_removes_certificate_from_iam,
 )
@@ -22,38 +22,62 @@ from tests.lib.update import (
 def subtest_update_happy_path(
     client, dns, tasks, route53, iam_govcloud, simple_regex, alb
 ):
-    operation_id = subtest_update_creates_update_operation(client, dns)
+    instance_model = ALBServiceInstance
+    operation_id = subtest_update_creates_update_operation(client, dns, instance_model)
     check_last_operation_description(client, "4321", operation_id, "Queuing tasks")
-    subtest_update_creates_private_key_and_csr(tasks)
-    subtest_gets_new_challenges(tasks)
-    subtest_update_updates_TXT_records(tasks, route53)
-    subtest_waits_for_dns_changes(tasks, route53)
-    subtest_update_answers_challenges(tasks, dns)
-    subtest_update_retrieves_new_cert(tasks)
-    subtest_update_uploads_new_cert(tasks, iam_govcloud, simple_regex)
-    subtest_update_selects_alb(tasks, alb)
-    subtest_update_adds_certificate_to_alb(tasks, alb)
-    subtest_update_provisions_ALIAS_records(tasks, route53, alb)
-    subtest_waits_for_dns_changes(tasks, route53)
+    subtest_update_creates_private_key_and_csr(tasks, instance_model)
+    subtest_gets_new_challenges(tasks, instance_model)
+    subtest_update_updates_TXT_records(tasks, route53, instance_model)
+    subtest_waits_for_dns_changes(tasks, route53, instance_model)
+    subtest_update_answers_challenges(tasks, dns, instance_model)
+    subtest_update_retrieves_new_cert(tasks, instance_model)
+    subtest_update_uploads_new_cert(tasks, iam_govcloud, simple_regex, instance_model)
+    subtest_update_selects_alb(tasks, alb, instance_model)
+    subtest_update_adds_certificate_to_alb(tasks, alb, instance_model)
+    subtest_update_provisions_ALIAS_records(tasks, route53, instance_model)
+    subtest_waits_for_dns_changes(tasks, route53, instance_model)
     subtest_update_removes_certificate_from_alb(tasks, alb)
-    subtest_update_removes_certificate_from_iam(tasks, iam_govcloud)
-    subtest_update_marks_update_complete(tasks)
+    subtest_update_removes_certificate_from_iam(tasks, iam_govcloud, instance_model)
+    subtest_update_marks_update_complete(tasks, instance_model)
 
 
-def subtest_update_selects_alb(tasks, alb):
+def subtest_update_creates_update_operation(client, dns, instance_model):
+    dns.add_cname("_acme-challenge.foo.com")
+    dns.add_cname("_acme-challenge.bar.com")
+    client.update_alb_instance("4321", params={"domains": "bar.com, Foo.com"})
+    db.session.expunge_all()
+
+    assert client.response.status_code == 202, client.response.body
+    assert "operation" in client.response.json
+
+    operation_id = client.response.json["operation"]
+    operation = db.session.get(Operation, operation_id)
+
+    assert operation is not None
+    assert operation.state == "in progress"
+    assert operation.action == "Update"
+    assert operation.service_instance_id == "4321"
+
+    instance = db.session.get(instance_model, operation.service_instance_id)
+    assert instance is not None
+    assert instance.domain_names == ["bar.com", "foo.com"]
+    return operation_id
+
+
+def subtest_update_selects_alb(tasks, alb, instance_model):
     db.session.expunge_all()
     alb.expect_get_certificates_for_listener("listener-arn-0", 1)
     alb.expect_get_certificates_for_listener("listener-arn-1", 5)
     alb.expect_get_listeners("listener-arn-0")
     tasks.run_queued_tasks_and_enqueue_dependents()
     alb.assert_no_pending_responses()
-    service_instance = db.session.get(ALBServiceInstance, "4321")
+    service_instance = db.session.get(instance_model, "4321")
     assert service_instance.alb_arn.startswith("alb-listener-arn-0")
 
 
-def subtest_update_adds_certificate_to_alb(tasks, alb):
+def subtest_update_adds_certificate_to_alb(tasks, alb, instance_model):
     db.session.expunge_all()
-    service_instance = db.session.get(ALBServiceInstance, "4321")
+    service_instance = db.session.get(instance_model, "4321")
     certificate = service_instance.new_certificate
     id_ = certificate.id
     alb.expect_add_certificate_to_listener(
@@ -63,16 +87,16 @@ def subtest_update_adds_certificate_to_alb(tasks, alb):
     tasks.run_queued_tasks_and_enqueue_dependents()
     alb.assert_no_pending_responses()
     db.session.expunge_all()
-    service_instance = db.session.get(ALBServiceInstance, "4321")
+    service_instance = db.session.get(instance_model, "4321")
 
     assert service_instance.new_certificate is None
     assert service_instance.current_certificate is not None
     assert service_instance.current_certificate.id == id_
 
 
-def subtest_update_provisions_ALIAS_records(tasks, route53, alb):
+def subtest_update_provisions_ALIAS_records(tasks, route53, instance_model):
     db.session.expunge_all()
-    service_instance = db.session.get(ALBServiceInstance, "4321")
+    service_instance = db.session.get(instance_model, "4321")
     example_com_change_id = route53.expect_create_ALIAS_and_return_change_id(
         "bar.com.domains.cloud.test", "alb.cloud.test", "ALBHOSTEDZONEID"
     )
@@ -80,15 +104,6 @@ def subtest_update_provisions_ALIAS_records(tasks, route53, alb):
         "foo.com.domains.cloud.test", "alb.cloud.test", "ALBHOSTEDZONEID"
     )
     tasks.run_queued_tasks_and_enqueue_dependents()
-
-
-def subtest_update_marks_update_complete(tasks):
-    tasks.run_queued_tasks_and_enqueue_dependents()
-    db.session.expunge_all()
-    service_instance = db.session.get(ALBServiceInstance, "4321")
-    operation = service_instance.operations.first()
-    assert operation
-    assert "succeeded" == operation.state
 
 
 def subtest_update_removes_certificate_from_alb(tasks, alb):
@@ -100,6 +115,34 @@ def subtest_update_removes_certificate_from_alb(tasks, alb):
     tasks.run_queued_tasks_and_enqueue_dependents()
 
     alb.assert_no_pending_responses()
+
+
+def subtest_update_uploads_new_cert(tasks, iam_govcloud, simple_regex, instance_model):
+    db.session.expunge_all()
+    service_instance = db.session.get(instance_model, "4321")
+    certificate = service_instance.new_certificate
+    today = date.today().isoformat()
+    assert today == simple_regex(r"^\d\d\d\d-\d\d-\d\d$")
+
+    iam_govcloud.expect_upload_server_certificate(
+        name=f"{service_instance.id}-{today}-{certificate.id}",
+        cert=certificate.leaf_pem,
+        private_key=certificate.private_key_pem,
+        chain=certificate.fullchain_pem,
+        path="/alb/external-domains-test/",
+    )
+
+    tasks.run_queued_tasks_and_enqueue_dependents()
+
+    db.session.expunge_all()
+    service_instance = db.session.get(instance_model, "4321")
+    certificate = service_instance.new_certificate
+    assert certificate.iam_server_certificate_name
+    assert certificate.iam_server_certificate_name.startswith("4321")
+    assert certificate.iam_server_certificate_id
+    assert certificate.iam_server_certificate_id.startswith("FAKE_CERT_ID")
+    assert certificate.iam_server_certificate_arn
+    assert certificate.iam_server_certificate_arn.startswith("arn:aws:iam")
 
 
 def subtest_update_noop(client):
