+++ tests/lib/update.py
@@ -22,42 +22,6 @@ def subtest_update_creates_private_key_and_csr(tasks, instance_model):
     )
 
 
-def subtest_update_creates_update_operation(client, dns, instance_model):
-    dns.add_cname("_acme-challenge.foo.com")
-    dns.add_cname("_acme-challenge.bar.com")
-    client.update_instance(
-        instance_model,
-        "4321",
-        params={
-            "domains": "bar.com, Foo.com",
-            "origin": "new-origin.com",
-            "path": "/somewhere-else",
-            "forward_cookies": "mycookie,myothercookie, anewcookie",
-            "forward_headers": "x-my-header, x-your-header   ",
-            "insecure_origin": True,
-        },
-    )
-    db.session.expunge_all()
-
-    assert client.response.status_code == 202, client.response.body
-    assert "operation" in client.response.json
-
-    operation_id = client.response.json["operation"]
-    operation = db.session.get(Operation, operation_id)
-
-    assert operation is not None
-    assert operation.state == "in progress"
-    assert operation.action == "Update"
-    assert operation.service_instance_id == "4321"
-
-    instance = db.session.get(instance_model, operation.service_instance_id)
-    assert instance is not None
-    assert instance.domain_names == ["bar.com", "foo.com"]
-    assert instance.cloudfront_origin_hostname == "new-origin.com"
-    assert instance.cloudfront_origin_path == "/somewhere-else"
-    return operation_id
-
-
 def subtest_gets_new_challenges(tasks, instance_model):
     db.session.expunge_all()
     tasks.run_queued_tasks_and_enqueue_dependents()
@@ -146,36 +110,6 @@ def subtest_update_retrieves_new_cert(tasks, instance_model):
     assert certificate.expires_at is not None
 
 
-def subtest_update_uploads_new_cert(
-    tasks, iam_commercial, simple_regex, instance_model
-):
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    certificate = service_instance.new_certificate
-    today = date.today().isoformat()
-    assert today == simple_regex(r"^\d\d\d\d-\d\d-\d\d$")
-
-    iam_commercial.expect_upload_server_certificate(
-        name=f"{service_instance.id}-{today}-{certificate.id}",
-        cert=certificate.leaf_pem,
-        private_key=certificate.private_key_pem,
-        chain=certificate.fullchain_pem,
-        path="/cloudfront/external-domains-test/",
-    )
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    db.session.expunge_all()
-    service_instance = db.session.get(instance_model, "4321")
-    certificate = service_instance.new_certificate
-    assert certificate.iam_server_certificate_name
-    assert certificate.iam_server_certificate_name.startswith("4321")
-    assert certificate.iam_server_certificate_id
-    assert certificate.iam_server_certificate_id.startswith("FAKE_CERT_ID")
-    assert certificate.iam_server_certificate_arn
-    assert certificate.iam_server_certificate_arn.startswith("arn:aws:iam")
-
-
 def subtest_update_marks_update_complete(tasks, instance_model):
     tasks.run_queued_tasks_and_enqueue_dependents()
     db.session.expunge_all()
@@ -185,14 +119,12 @@ def subtest_update_marks_update_complete(tasks, instance_model):
     assert "succeeded" == operation.state
 
 
-def subtest_update_removes_certificate_from_iam(tasks, iam_commercial, instance_model):
-    iam_commercial.expects_delete_server_certificate(
-        f"4321-{date.today().isoformat()}-1"
-    )
+def subtest_update_removes_certificate_from_iam(tasks, iam, instance_model):
+    iam.expects_delete_server_certificate(f"4321-{date.today().isoformat()}-1")
 
     tasks.run_queued_tasks_and_enqueue_dependents()
 
-    iam_commercial.assert_no_pending_responses()
+    iam.assert_no_pending_responses()
     instance = db.session.get(instance_model, "4321")
     assert len(instance.certificates) == 1
